fix profile for gnome-xterm
this file was executed by the the standard Bourne/POSIX/Korn
shell as a login shell. For bash, this file was executed if
.bash_profile or .bash_login did not exist.

See /usr/share/doc/bash/examples/startup-files for examples.

For GDM the X Window Manager in Ubuntu, it would read this file.
Please see /etc/gdm3/Xsession.

We linked .profile to .bash_profile to sync the all PATH setup